requirements.txt
# Python 3.10+ Data Engineering Project Requirements

# Financial data retrieval
yfinance>=0.2.40

# Apache Kafka integration
kafka-python>=2.0.2

# Apache Spark for distributed data processing
pyspark>=3.5.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.9

# Data manipulation and analysis
pandas>=2.2.0

# Environment variable management
python-dotenv>=1.0.0

# Code linting and formatting
ruff>=0.3.0
